requirements.txt
cloudshell-automation-api>=9.3,<2020.1